Extract editable cleanup support
