+++ src/agilab/core/agi-cluster/src/agi_cluster/agi_distributor/deployment_editable_install_support.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 import hashlib
 import json
 import os
+import shutil
 from pathlib import Path
 from typing import Any
 from urllib.parse import unquote, urlparse
@@ -20,10 +21,79 @@ from agi_cluster.agi_distributor.deployment_venv_support import (
 )
 
 EDITABLE_INSTALL_CACHE_SCHEMA = "agilab-editable-install-cache-v1"
+EDITABLE_SHADOW_IMPORTS: dict[str, tuple[str, ...]] = {
+    "agi-cluster": ("agi_cluster",),
+    "agi-core": ("agi_core",),
+    "agi-env": ("agi_env",),
+    "agi-node": ("agi_node",),
+    "agilab": ("agilab",),
+}
 
 
 def _is_python_project(path: Path) -> bool:
     return (path / "pyproject.toml").exists() or (path / "setup.py").exists()
+
+
+def _cleanup_editable(site_packages: Path) -> None:
+    patterns = (
+        "__editable__.agi_env*.pth",
+        "__editable__.agi_node*.pth",
+        "__editable__.agi_core*.pth",
+        "__editable__.agi_cluster*.pth",
+        "__editable__.agilab*.pth",
+    )
+    for pattern in patterns:
+        for editable in site_packages.glob(pattern):
+            try:
+                editable.unlink()
+            except FileNotFoundError:
+                pass
+
+
+def _editable_shadow_import_names(package_project: Path) -> tuple[str, ...]:
+    distribution_name = package_project.name.replace("_", "-").lower()
+    return EDITABLE_SHADOW_IMPORTS.get(distribution_name, ())
+
+
+def _remove_site_package_shadow(path: Path) -> bool:
+    try:
+        if path.is_dir() and not path.is_symlink():
+            shutil.rmtree(path)
+            return True
+        if path.exists() or path.is_symlink():
+            path.unlink()
+            return True
+    except FileNotFoundError:
+        return False
+    return False
+
+
+def _cleanup_editable_shadow_packages(
+    venv_project: Path,
+    package_projects: list[Path],
+    *,
+    os_name: str = os.name,
+    python_version: str | None = None,
+) -> list[Path]:
+    """Remove stale package dirs that shadow editable source installs."""
+    site_packages = _project_site_packages_dir(
+        venv_project,
+        os_name=os_name,
+        python_version=python_version,
+    )
+    if not site_packages.exists():
+        return []
+
+    removed: list[Path] = []
+    for package_project in package_projects:
+        for import_name in _editable_shadow_import_names(package_project):
+            for candidate in (
+                site_packages / import_name,
+                site_packages / f"{import_name}.py",
+            ):
+                if _remove_site_package_shadow(candidate):
+                    removed.append(candidate)
+    return removed
 
 
 def _editable_install_cache_path(venv_project: Path) -> Path:
